+++ examples/langgraph/data/simple_prompts.jsonl
@@ -1,3 +1,1 @@
 {"name":"p1","prompt":"Say hello in one sentence","gt":"hello"}
-{"name":"p2","prompt":"Introduce yourself briefly","gt":"intro"}
-{"name":"p3","prompt":"Respond with a fun fact about space","gt":"space"}
